add macros for item translation (#237)
+ Macro for storing published/corrected article

This macro is meant for use with articles that have been auto translated to nynorsk from bokmål. In most cases reporters will make corrections since the auto translation is not perfect. This macro saves the corrected/published version in the translator's database. (Will be used in the future for a kind of correction middleware)