Macro changes (#247)
Added logging, etc., and changed access_type to frontend.

* Macro for storing published/corrected article

This macro is meant for use with articles that have been auto translated to nynorsk from bokmål. In most cases reporters will make corrections since the auto translation is not perfect. This macro saves the corrected/published version in the translator's database. (Will be used in the future for a kind of correction middleware)

* Reducing payload

I originally sent over the entire item data structure, manipulated some of the elements on the server side and sent it back. Somehow that corrupted a few of the fields I *didn't* do anything with. So now I've changed the code to just send over the relevant fields.

* Reduce payload

Strictly speaking, the change I've made here isn't necessary. However to keep this in sync and in style with the nob_NO_translate macro, I've done it anyway. The difference is that with this change I just send over the necessary fields instead of the whole data structure.