[FEED PARSER] [WUFOO] Don't fail on missing field
if a field is missing, an empty string is used instead.
In the case of "birth date", a warning is printed and current date is
used.

This patch also fixes code style in few files and disable W503 warning
for flake8, which is not valid anymore (PEP 8 has changed on this point).

SDNTB-509