[FEED PARSER][SOLITA][FORMATTER][NTB NITF]add solita images to extra + output them in NITF + genre fix
image are added to a private data in item['extra']['ntb_media'], and are
formatted in NTB NITF. The solita parser now uses a "ntb-media" class
to mark the paragraph where image link are put, and this is used by NTB
NITF to remove it from body.

Extra data are added when available in <media> element and its child
(`class` and `alternate-text`)

Fixed genre.

SDNTB-561
SDNTB-566